==> Tree_Analysis/Step2_Tree_Carbon_Estimates.R <==
path_to_tree_folder = "C:/Users/Chels/OneDrive - University of Illinois - Urbana/Ch2_Floodplain_Experiment/Floodplain-Experiment-Repo/Tree_Analysis"
setwd(path_to_tree_folder)

library(allodb)
library(dplyr)
library(tidyr)
library(reshape2)

# conversions
m2.per.ha = 10^4
kg.per.Mg = 10^3

# constants
tree.C.content = 0.48 # %

# plot dimensions 
plot.length = 48 #m 
plot.width = 12 #m 
plot.area.m2 = plot.length * plot.width #m2
plot.area.ha = plot.area.m2 / m2.per.ha #ha

################################################################################
# step 1: estimate above- and belowground biomass from diameter-at-breast height (DBH) data (cm) 

# define data-related variables
trt.names = c("Balled-and-burlapped","Bareroot","Seedling","Acorn","Seedbank","Reference")
years = c(2022, 2023)
n.y = length(years)

# list of genus families
family.df = read.csv("Tree_Databases/Tree_Families.csv")
n.f = nrow(family.df)
genus.families = list()
for (i in 1:n.f) { genus.families[[family.df$Genus[i]]] = family.df$Family[i] }

# read in all DBH data
dbh.all = read.csv("Clean_Data_By_Species/DBH_Data_Clean_2022_2023.csv", header=T)

# make list for unique species
uni.spp = sort(unique(dbh.all$spp))
n.sp = length(uni.spp)

# read in allometric equation matrices
allo.df1 = read.csv("Tree_Databases/Jenkins2004.csv", header=T); rownames(allo.df1) = allo.df1$spp
allo.df2 = read.csv("Tree_Databases/Chojnacky2014.csv", header=T); rownames(allo.df2) = allo.df2$spp

# estimate biomass with allometric eqns in 1: Jenkins (2004) & 2: Chojnacky (2014)
new.mat = data.frame(matrix(nrow = length(dbh.all$plot), ncol = 6))
colnames(new.mat) = c("ab1","ab2","r1","r2","bg1","bg2")
dbh.all = cbind(dbh.all, new.mat)
for (sp in uni.spp) {
  sp.ind = which(dbh.all$spp == sp)
  
  # Jenkins
  dbh.all[sp.ind,"ab1"] = exp(allo.df1[sp,"ab.b0"] + allo.df1[sp,"ab.b1"]*log(dbh.all[sp.ind,"dbh.cm"]))/kg.per.Mg # bm = exp(b0 + b1*log(dbh [cm])) [kg -> Mg]
  dbh.all[sp.ind,"r1"] = exp(allo.df1[sp,"cr.b0"] + allo.df1[sp,"cr.b1"]/dbh.all[sp.ind,"dbh.cm"]) # ratio = exp(b0 + b1/(dbh [cm]))
  dbh.all[sp.ind,"bg1"] = dbh.all[sp.ind,"r1"]*dbh.all[sp.ind,"ab1"] 
  
  # Chojnacky
  dbh.all[sp.ind,"ab2"] = exp(allo.df2[sp,"ab.b0"] + allo.df2[sp,"ab.b1"]*log(dbh.all[sp.ind,"dbh.cm"]))/kg.per.Mg # bm = exp(b0 + b1*log(dbh [cm])) [kg -> Mg]
  dbh.all[sp.ind,"r2"] = exp(allo.df2[sp,"cr.b0"] + allo.df2[sp,"cr.b1"]*log(dbh.all[sp.ind,"dbh.cm"])) # ratio = exp(b0 + b1*log(dbh [cm]))
  dbh.all[sp.ind,"bg2"] = dbh.all[sp.ind,"r2"]*dbh.all[sp.ind,"ab2"]
}

# estimate biomass with allodb
dbh.all$ab3 = get_biomass(dbh = dbh.all$dbh.cm, 
                          genus = dbh.all$genus, 
                          species = dbh.all$species, 
                          coords = c(-90.1835, 41.5542))/kg.per.Mg # [kg -> Mg]
dbh.all$bg3 = dbh.all$ab3 * dbh.all$r2 # using Chojnacky root:shoot ratio
mean(dbh.all$r2)

# check for NAs
sum(is.na(dbh.all[,c("ab1","ab2","r1","r2","bg1","bg2","ab3","bg3")]))

# sum above & belowground biomass by species within each plot
biomass.by.species = dbh.all %>% 
                     group_by(redo, year, treatment, full.treatment.name, 
                              plot, spp, genus, species) %>% 
                     summarise(ab1=sum(ab1), ab2=sum(ab2), ab3=sum(ab3), 
                               bg1=sum(bg1), bg2=sum(bg2), bg3=sum(bg3))

# divide biomass by total plot area (Mg/ha)
bm.vars = c("ab1","ab2","ab3","bg1","bg2","bg3")
bm.vars.ha = c("ab.ha1","ab.ha2","ab.ha3","bg.ha1","bg.ha2","bg.ha3")
biomass.by.species[, bm.vars] = biomass.by.species[, bm.vars]/plot.area.ha
colnames(biomass.by.species)[which(colnames(biomass.by.species) %in% bm.vars)] = bm.vars.ha

################################################################################
# step 2: estimate carbon stocks from biomass (Mg C/ha)

# clean wood database
wood.c.df = read.csv("Tree_Databases/Doraisami_2021_Wood_C_Database.csv")
live.wood.c.df = wood.c.df[which(wood.c.df$dead.alive == "alive" & wood.c.df$growth.form == "tree"),]
live.stem.c.df = live.wood.c.df[which(live.wood.c.df$tissue == "stem"),]
live.stem.c.df = live.stem.c.df %>% separate(binomial.resolved, c("genus","spp"), sep="_", remove=F)
live.stem.c.df$species = paste(live.stem.c.df$genus, live.stem.c.df$spp, sep= " ")

# get best-available wood C concentration for each species
biomass.by.species$taxon.c.content = rep(0, nrow(biomass.by.species))
biomass.by.species$binomial = paste(biomass.by.species$genus, biomass.by.species$species, sep=" ")
biomass.by.species$family = rep(NA, nrow(biomass.by.species))
for (i in 1:nrow(biomass.by.species)) {
  if (biomass.by.species$binomial[i] != "") {
    biomass.by.species$family[i] = genus.families[[biomass.by.species$genus[i]]]
    df.sp.id = which(live.stem.c.df$species == biomass.by.species$binomial[i])
    df.gen.id = which(live.stem.c.df$genus.resolved == biomass.by.species$genus[i]) 
    df.fam.id = which(live.stem.c.df$family.resolved == biomass.by.species$family[i]) 
    if (length(df.sp.id) > 0) {
      biomass.by.species[i,"taxon.c.content"] = mean(live.stem.c.df[df.sp.id,"tissue.c"])/100
    } else if (length(df.gen.id > 0)) {
      biomass.by.species[i,"taxon.c.content"] = mean(live.stem.c.df[df.gen.id,"tissue.c"])/100
    } else if (length(df.fam.id > 0)) {
      biomass.by.species[i,"taxon.c.content"] = mean(live.stem.c.df[df.fam.id,"tissue.c"])/100
    } else {
      biomass.by.species[i,"taxon.c.content"] = tree.C.content
    }
  } else {
    biomass.by.species[i,"taxon.c.content"] = tree.C.content
  }
}

# estimate C stock by species
id.vars = c("redo","year","treatment","full.treatment.name","plot","spp","genus","species")
C.vars.ha = c("abC.ha1","abC.ha2","abC.ha3","bgC.ha1","bgC.ha2","bgC.ha3")
C.stocks.by.species = biomass.by.species[, c(id.vars, bm.vars.ha)]
C.stocks.by.species[,bm.vars.ha] = biomass.by.species[,bm.vars.ha] * biomass.by.species$taxon.c.content
colnames(C.stocks.by.species)[which(colnames(C.stocks.by.species) %in% bm.vars.ha)] = C.vars.ha

# sum C stocks by plot (Mg/ha)
C.stocks.by.plot <- C.stocks.by.species %>% 
                    group_by(redo, year, treatment, full.treatment.name, plot) %>% 
                    summarise(abC.ha1=sum(abC.ha1), 
                              abC.ha2=sum(abC.ha2), 
                              abC.ha3=sum(abC.ha3), 
                              bgC.ha1=sum(bgC.ha1), 
                              bgC.ha2=sum(bgC.ha2), 
                              bgC.ha3=sum(bgC.ha3))

# estimate total biomass C stocks per ha
C.stocks.by.plot = C.stocks.by.plot %>% 
                   group_by(redo, year, treatment, full.treatment.name, plot) %>% 
                   summarise(abC.ha1 = abC.ha1, 
                             abC.ha2 = abC.ha2, 
                             abC.ha3 = abC.ha3,
                             bgC.ha1 = bgC.ha1, 
                             bgC.ha2 = bgC.ha2, 
                             bgC.ha3 = bgC.ha3,
                             bmC.ha1 = abC.ha1 + bgC.ha1,
                             bmC.ha2 = abC.ha2 + bgC.ha2,
                             bmC.ha3 = abC.ha3 + bgC.ha3)

################################################################################
# step 3: replace bad 2022 data with 2023 redo data

# redo plots from 2023
redo.ind = which(C.stocks.by.plot$redo == "Y")
redo.plot.df = C.stocks.by.plot[redo.ind,]
redo.plot.df$treatment_plot = paste(redo.plot.df$treatment, redo.plot.df$plot, sep="")
redo.trt.plots = unique(redo.plot.df$treatment_plot)

# original plots from 2022
C.stocks.by.plot$treatment_plot = paste(C.stocks.by.plot$treatment, C.stocks.by.plot$plot, sep="")
orig.ind = which((C.stocks.by.plot$redo == "N" & C.stocks.by.plot$year == 2022) & (C.stocks.by.plot$treatment_plot %in% redo.trt.plots))
C.stocks.by.plot[orig.ind,"treatment_plot"]

# remove C stocks from 2022 that were re-done in 2023 by plot
C.stocks.by.plot.redo = C.stocks.by.plot[-orig.ind,]
C.stocks.by.plot.redo[which(C.stocks.by.plot.redo$year == 2023),"year"] = 2022

# remove C stocks from 2022 that were re-done in 2023 by species
C.stocks.by.species$treatment_plot = paste(C.stocks.by.species$treatment, C.stocks.by.species$plot, sep="")
orig.ind.spp = which((C.stocks.by.species$redo == "N" & C.stocks.by.species$year == 2022) & (C.stocks.by.species$treatment_plot %in% redo.trt.plots))
C.stocks.by.species.redo = C.stocks.by.species[-orig.ind.spp,]
C.stocks.by.species.redo[which(C.stocks.by.species.redo$year == 2023),"year"] = 2022

# estimate total biomass C stocks per ha by species
C.stocks.by.species.redo = C.stocks.by.species.redo %>% 
                           group_by(redo, year, treatment, full.treatment.name, 
                                    plot, spp, genus, species) %>%
                           summarise(abC.ha1 = abC.ha1, 
                                     abC.ha2 = abC.ha2, 
                                     abC.ha3 = abC.ha3,
                                     bgC.ha1 = bgC.ha1, 
                                     bgC.ha2 = bgC.ha2, 
                                     bgC.ha3 = bgC.ha3,
                                     bmC.ha1 = abC.ha1 + bgC.ha1,
                                     bmC.ha2 = abC.ha2 + bgC.ha2,
                                     bmC.ha3 = abC.ha3 + bgC.ha3)

# write C stocks results to file
write.csv(C.stocks.by.plot.redo, "Clean_Data_By_Plot/WoodyBiomass_C_Stocks_By_Plot.csv", row.names=F)
write.csv(C.stocks.by.species.redo, "Clean_Data_By_Species/WoodyBiomass_C_Stocks_By_Species.csv", row.names=F)
